depre: Rename `QEInputMaker` to `FromQEOutput`
The old name of the class, `QEInputMaker`, sounds like we
are making some input for Quantum ESPRESSO, but it is actually
not the case. In fact, we are using Quantum ESPRESSO's output
as input to make the real "input" for `qha`. This change will
make the class more "as is".
Breaking: If you have every used this class in your own code!