depre: Rename 'make_input.py' to 'input_maker.py'
Why? Because I found that we put both runnable scripts, like
'plot.py', 'run.py' and 'convert.py' together with other standalone modules,
so I decide: For runnable scripts, we use a verb to name it,
for others, we use a noun to name them. But 'make_input.py'
is not a runnable script, so I made this rename. And it is not suitable
to put it in 'readers' sub package, so I moved it out.